made few changes.

+++ sbpc.Timesheet.Data/Entity/Job.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public decimal OverTimeRate { get; set; }
+        public double OverTimeRate { get; set; }
         public double CostPerMile { get; set; }
         public bool Active { get; set; }
     }
